+++ src/android/com/pbakondy/Sim.java
@@ -209,9 +209,15 @@ public class Sim extends CordovaPlugin {
         result.put("subscriberId", subscriberId);
       }
 
+      // android.telephony.SubscriptionManager#getDefaultDataSubscriptionId requires API 24
+      if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
+        result.put("defaultDataSubscriptionId", SubscriptionManager.getDefaultDataSubscriptionId());
+      }
+      
       if (sims != null && sims.length() != 0) {
         result.put("cards", sims);
       }
+
 
       callbackContext.success(result);